requirements.txt
cachetools==5.0.0
certifi==2021.10.8
charset-normalizer==2.0.12
google-api-core==2.7.2
google-api-python-client==2.46.0
google-auth==2.6.6
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.5.1
googleapis-common-protos==1.56.0
httplib2==0.20.4
idna==3.3
oauthlib==3.2.0
protobuf==3.20.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==3.0.8
python-dotenv==0.20.0
requests==2.27.1
requests-oauthlib==1.3.1
rsa==4.8
six==1.16.0
tabulate==0.8.9
uritemplate==4.1.1
urllib3==1.26.9
